FINE_official/requirements.txt
# torchvision==0.9.0
requests==2.23.0
tqdm==4.28.1
# torch==1.8.0
# numpy==1.19.5
mlflow==1.11.0
Pillow==8.1..0
# scikit_learn==0.24.0
wandb
torchnet